correcoes fk em delete usecases

+++ src/main/java/jonatasSantos/royalLux/core/application/usecases/material/MaterialDeleteUseCaseImpl.java
@@ -2,6 +2,7 @@ package jonatasSantos.royalLux.core.application.usecases.material;
 
 import jakarta.persistence.EntityNotFoundException;
 import jonatasSantos.royalLux.core.application.contracts.repositories.MaterialRepository;
+import jonatasSantos.royalLux.core.application.contracts.repositories.MaterialSalonServiceRepository;
 import jonatasSantos.royalLux.core.application.contracts.usecases.material.MaterialDeleteUseCase;
 import jonatasSantos.royalLux.core.application.models.dtos.material.MaterialDeleteUseCaseOutputDto;
 import org.springframework.stereotype.Service;
@@ -10,15 +11,22 @@ import org.springframework.stereotype.Service;
 public class MaterialDeleteUseCaseImpl implements MaterialDeleteUseCase {
 
     private final MaterialRepository materialRepository;
+    private final MaterialSalonServiceRepository materialSalonServiceRepository;
 
-    public MaterialDeleteUseCaseImpl(MaterialRepository materialRepository) {
+    public MaterialDeleteUseCaseImpl(MaterialRepository materialRepository, MaterialSalonServiceRepository materialSalonServiceRepository) {
         this.materialRepository = materialRepository;
+        this.materialSalonServiceRepository = materialSalonServiceRepository;
     }
 
     @Override
     public MaterialDeleteUseCaseOutputDto execute(Integer id) {
         var material = this.materialRepository.findById(id.toString())
                 .orElseThrow(() -> new EntityNotFoundException("Material inexistente"));
+
+        var materialsSalonServices = this.materialSalonServiceRepository.findByMaterialId(material.getId());
+
+        if (!materialsSalonServices.isEmpty())
+            throw new IllegalStateException("Material não pode ser deletado pois ainda possui serviços vinculados");
 
         this.materialRepository.delete(material);
